requirements.txt
python-telegram-bot==20.3
gspread==5.9.0
oauth2client==4.1.3
flask==2.2.5

